Moves plugin-specific code to plugins module
Quite a bit of the code migrated in from SQLAlchemy-Utils only gets used
within the plugins module. There's no need for us to incur the
initialization expense if those classes and methods aren't needed or
used.

This will only actually have an initialization benefit if the plugins
module does not load every plugin upon import. However, having the code
co-located anyway makes sense.
